requirements.txt
feedparser>=6.0.0
markdownify>=0.11.0
requests>=2.28.0
python-slugify>=8.0.0
